schema: add s390x as a valid linux architecture
This adds a new `s390x` architecture label for linux.
It maps to IBM z/Architecture: https://en.wikipedia.org/wiki/Z/Architecture

`s390x` is a 64-bits, alignes access, big-endian architecture:
https://wiki.debian.org/ArchitectureSpecificsMemo

This new label corresponds to `uname -m` on Linux systems, to GOARCH and
to the usual distro naming.
The usual GNU triplet for this is `s390x-ibm-linux-gnu`.