remove k8s.io/kubernetes dependency
This commit removes k8s.io/kubernetes as a dependency. The one part of
the repo that we were using, k8s.io/kubernetes/pkg/api/resource, has
been copied to schema/types/resource, and very minimally modified to
work in its new location. A README.md has been added to this new
location describing where the files came from.

This was necessary to avoid a cyclic dependency with the kubernetes
repo.