schema/labels: consistently order LabelsFromMap
Prior to this change, the slice of labels produced by LabelsFromMap
could have any arbitrary ordering.

See https://github.com/appc/docker2aci/issues/245 for one of the issues
this causes in practice.

I also renamed the type 'labels' to 'labelsSlice' since, as it was, it
was being constantly shadowed.

This is not accompanied with any spec change since labels in the appc
spec are already a slice, and thus already have a consistent ordering.
It's unfortunate that a round trip of ToMap/FromMap won't be consistent,
but I don't think any callers actually do that currently.